requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
python-jose
passlib[bcrypt]
argon2-cffi
python-dotenv
pydantic_settings